Bug-1701: Fix type overflow
Corrected a bug that caused Movie IDs to be read as int instead of long, resulting in Movies not being returned if their ID was higher than the int max value.

Created a unit test for the overflow and fixed an existing unit test that was failing due to a bad constant.